fix: split components into more files
