+++ src/lib/Priority.ts
@@ -1,0 +1,5 @@
+export enum Priority {
+	LOW = 0,
+	NORMAL = 1,
+	HIGH = 2
+}
